Create KeyCodes and MouseCodes

+++ PepperMint/src/PepperMint/Debug/Instrumentor.h
@@ -11,7 +11,7 @@ namespace PepperMint {
 struct ProfileResult {
 	const char* name;
 	long long start, end;
-	uint32_t threadId;
+	size_t threadId;
 };
 
 struct InstrumentationSession {
@@ -99,7 +99,7 @@ private:
 		long long start = std::chrono::time_point_cast<std::chrono::microseconds>(_startTimepoint).time_since_epoch().count();
 		long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-		uint32_t threadId = std::hash<std::thread::id>{}(std::this_thread::get_id());
+		size_t threadId = std::hash<std::thread::id>{}(std::this_thread::get_id());
 		Instrumentor::Get().writeProfile({ _name, start, end, threadId });
 
 		_stopped = true;
